test: handle test case for updated datasource manager
